Multi-OS path issues corrected
The path separators are back to being replaced by the current OS ones
regardless of the type of path being used so that they're ready to use
in case they're supported by the OS (a relative path always is - unlike
an absolute one, which may or may not be).

Also before any attempt to read/write path stuff, the path is checked if
it's supported by the current OS (even though /srv/something is
supported on Windows and maps to C:\srv\something, I'm supposing that's
Linux-only to stop Windows from trying to read and/or write to it as it
has already done. And I've put the same check on Linux:
[Letter]:\something is not supported there, nor \\something is (Windows
network path)).

Also now the modules print their name before a critical error right at
the beginning with the personal constants.